+++ backend/src/main/java/com/rungo/api/domain/marathon/marathon/dto/CreateMarathonRes.java
@@ -2,6 +2,7 @@ package com.rungo.api.domain.marathon.marathon.dto;
 
 import com.rungo.api.domain.marathon.marathon.enumtype.MarathonStatus;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -22,7 +23,7 @@ public record CreateMarathonRes(
     public record CourseItemRes(
             Long id,
             String courseType,
-            Integer price,
+            BigDecimal price,
             Integer capacity,
             Integer currentCount,
             Integer remainingCount
